Allow to change modifiers for panning and zooming
For some applications, the defaults were inconvenient.
For example, sometimes it is better to pan by default
because users need it more often than zoom.

Also, switch to getModifiersEx() instead of getModifiers().

Compatibility warning: panMask is protected, so whatever
subclases expected to find old-style modifiers there
(getModifiers()) will fail to do so (because it is now
in getModifiersEx() language).